Adding in test dependencies.
The `multi` test is a test for stealjs being used by multi node modules at the same time. There was a bug that when this happened the second module using stealjs would throw an exception.

For this commit I had to change the .gitignore so that node_modules folders are committed. They are for the sake of the test, stealjs's node_modules are not checked in.