Removing use of __useDefault
Removing all uses of __useDefault in tests. We're no longer going to
support having the `default` export in es6 modules being imported into
non-es6 as the module's value. SystemJS no longer supports this and
hacking it back in would be unnecessarily difficult. There's no real
need for the feature anyways.