Override instantiate to ensure @config is executed first
This changes how `System.main` is run in production. Since we are now
going to be transpiling `@config` to AMD, we need a way to execute it
before anything else is defined. We can do this by overriding
System.instantiate and loading `@config` and it's dependencies before
anything else is instantiated.