In production, make @config part of main bundle
This changes the way we load config-first in production to get rid of the instantiate hook in startup and instead add `@config` to the main bundle and import it prior to main.

This will necessitate a change in steal-tools, the `@config` cannot be a global because it will throw here: https://github.com/ModuleLoader/es6-module-loader/blob/v0.9.4/src/loader.js#L882

The fix will be to convert `@config` modules to AMD in the build.

Should fix #275