Tilde extension to allow replacing plugin modules
This implements the tilde extension `src/system-extension-tilde.js` which makes it easy for plugins to define places that can be replaced by `normalized` and `located` modules.  The idea is that a plugin can identify a regular expression (or an array of regular expressions) which will be used to replace moduleNames located in those regular expressions with a full address which has been found by running the module name through normalize and locate hook.

This applies replacements for less and css as outlined in issue #277