Implement locate://, pkg:// and update test. Not ready to merge.
locate:// & pkg:// are both capable of importing packages at all locations, parent relative, baseURL relative & module paths, using the same syntax.

they differ in how they output the resolved path. locate:// outputs a path relative to the file doing the import, pkg:// outputs a path relative to the baseurl.

this is important in some cases, like less & css, as paths in imported files shouldn't be relative to the imported file location, they should be relative to the root importer location.

i'm not convinced pkg:// is required, or if it is it should be renamed. as the alternative locate:// when used in less and css should be aware of the root importer location and use that as the "base".

should have a way to disable that behaviour in case the user is using the less option that rewrites all the imports to be relative to the import parent instead of root.