Do not allow cloned loader to leak scope
When creating a new loader we need a completely new loader with no existing modules. Previously we were copying the existing modules over and deleting parent modules.

The problem with this method is that it means that modules share their scope with a cloned loader and the "real" loader. So anything that is modified within that scope within the cloned loader code... will modify the outside scope as well..

The solution is to no loader copy the existing loads. Rather since we are already retaining the source, that will be enough.