Fix progressively loaded plugins configuration
This fixes https://github.com/stealjs/steal/issues/957

When we progressively load a package we also load in `plugins`
configuration it has. This configuration should be treated not as from
the root package, because doing so has been causing it to be looked up
from too deep of a folder. This closes #957