Use resolutions to determine which pkg version to use.
This bug depends on precise loading timing that I was unable to recreate
in a test. What needs to happen is module A and B need to load the dep
at the exact same time, with module B getting an incorrect version the
first time, and module A using that incorrect version in its normalize
resolution.

This fixes #1075