Tighten restriction on use of devDependencies
This tightens when we consider a devDependency to be part of the
dependencyMap. This map is used for a few things in the npm plugin. We
only want to include devDependency in this map when trying to normalize
a name. This changes the method so that it accepts a boolean that
decides where devDependencies should be included or not, so that it is
restricted to just that use case.

Fixes #1086