Fixes incorrect baseURL in IE
This change mimics steal@0.16 behavior in browsers without
`document.currentScript` support; meaning, the last script tag on the
page is assumed to be the one loading `steal.js`

The previous approach of attaching a `onload` handler to the the
scripts fails in IE10 because `load` events are fired regardless the
script succeeded to load or not.

If multiple script tags happen to load scripts from paths that match
`/steal/` an incorrect `baseURL` value could be set.

Fixes #1094