Prevent mutating params
This fixes all instances where we mutate a function argument. The case where this is bad is when `arguments` are forwarded to another function. This causes a deopt in V8.

Since there is no linting rule to only fix instances where arguments are forwarded, we are just going to avoid mutating parameters altogether and instead assign a local variable. Closes #1157