Cache normalize call
This makes it so that the normalize call is cached based on the module identifier and the parent module name. The cached name is only used after a module has been imported, so that if a normalize applies to a module not yet imported, we don't accidentally cache an name not used for importing a module.

Closes #1158