Simplify mismatch version error message
Before it was printing out all the package names that depends on the
package that can't be found. That's confusing, specially since there can
be a lot of them.

The error message looks like this now:

Did not find [pkg.origFileUrl],
Unable to find a compatible version of [pkg.name]
Wanted: [pkg.version]
Found: [found.version]

Closes #1176