Prevent false positive 'loaded twice' warnings
This changes how the false positive warnings work, requiring you to
*explicitly* set the log level to `0`. This means you now need to do:

```html
<script src="./node_modules/steal/steal.js" log-level="0"></script>
```

To see these warnings. This was the only option I could come up with,
   given that this type of message is prone to false-positives (because
 steal's registry is keyed by moduleName, not address).

For 2.0 I want to fix this log level stuff. Currently it defaults to 0.
It should default to 'warn' level or maybe even error.

Fixes #1285