Prevent rerunning module loading on modules that have already failed
This is a bug fix deep within steal's core. The issue I was trying to
fix is #1317, where the 'module loaded twice' bug.

What I found was that when a module fails to load it gets removed from
`loader.loaderObj.loads`, an array that contains all of the loaded (or
		loading) loads.

Because it gets removed, when another module requests the same (failed)
	load, steal will start loading it again, under a new record.

This is the root cause of steal's weirdness where it will show the same
error message multiple times when working on large codebases.

So this fixes that, a pretty core issue, but also fixes #1317 in the
process.