Support tree-shaking graphs with multi-level re-export modules
This makes it possible for us to support tree-shaking in graphs where
there are multiple levels of re-exports. This is achieved by determining
used exports by walking up the graph, in a single `while` loop, and
tracking export names until we come to an `import`ing module, and count
those as used exports.