Set `less` to be async by default
A recent version of less now sets `display: none` during initialization.
The only way to prevent this is by setting `window.less = { async: false
	};`.

Because this needs to be a super dependancy it cannot be in steal-less.
The only solution, unfortunately, is for steal to set this.

Fixes https://github.com/stealjs/steal-less/issues/76