Prevent using babel-code-frame in older browsers
babel-code-frame is the module we use to pretty-print the location
within code where an Error was thrown. However internally (we use a
		bundle of the module) it depends on some code that depends on
`__proto__`. So browsers without support for that will throw when
codeFrame is used.

The result is that users will see an error about some internal steal bug
rather than the error from their code. Since we can't pretty-print using
babel-code-frame on these older browsers, this restores the default
behavior of just bubbling the original error.

Fixes #1466