Prevent running normalization on already normalized npm names
This prevents a bug caused by rollup in tree-shaking, where it will
reorder dependencies. Some times it might add a dependency to a module
that already shares that same dependency, but in a different version. In
this scenario the logic of the npm extension will look for the module's
version, and normalize incorrectly. Since the full npm name is already
encoded we can skip the expensive npm normalization logic altogether.
Fixes https://github.com/stealjs/steal/issues/1490