Update Saucelabs platform parameters
The tests are failing with "The environment you requested was
unavailable". In the past this was related to the appium version.

The parameters in this commit came from the saucelabs configurator,
I tried to introduce as little changes as possible, but some of the
old parameters were not available anymore (like the iOS version)